backend/requirements.txt
fastapi
uvicorn[standard]
celery
redis
python-dotenv
openai
pandas
openpyxl
python-multipart
aiofiles
flower